refactor!: remove Sysdig Secure tools, redirect to @sysdig/secure-mcp-server
BREAKING CHANGE: list_runtime_events, get_event_info and get_event_process_tree
have been removed from this server. Use the dedicated @sysdig/secure-mcp-server
npm package for Sysdig Secure workflows. SysQL tools (generate_sysql, run_sysql)
remain because they target both Monitor and Secure datasets.
