+++ README.md
@@ -2,6 +2,13 @@
 
 [![App Test](https://github.com/sysdiglabs/sysdig-mcp-server/actions/workflows/publish.yaml/badge.svg?branch=main)](https://github.com/sysdiglabs/sysdig-mcp-server/actions/workflows/publish.yaml)
 
+> [!IMPORTANT]
+> **Breaking change — this MCP server now focuses on Sysdig Monitor.**
+>
+> Starting with the next major release, the dedicated Sysdig Secure tools (`list_runtime_events`, `get_event_info`, `get_event_process_tree`) have been removed from this server. For Sysdig Secure use cases, install the new **[@sysdig/secure-mcp-server](https://www.npmjs.com/package/@sysdig/secure-mcp-server)** package, which provides comprehensive coverage of Sysdig Secure capabilities.
+>
+> The SysQL tools (`generate_sysql`, `run_sysql`) remain available here because they can be used against both Monitor and Secure datasets.
+
 ---
 
 ## Table of contents
@@ -28,14 +35,16 @@
 
 ## Description
 
-This is an implementation of an [MCP (Model Context Protocol) Server](https://modelcontextprotocol.io/quickstart/server) to allow different LLMs to query information from the Sysdig platform (Monitor and Secure). New tools and functionalities will be added over time following semantic versioning. The goal is to provide a simple and easy-to-use interface for querying information from the Sysdig platform using LLMs.
+This is an implementation of an [MCP (Model Context Protocol) Server](https://modelcontextprotocol.io/quickstart/server) that exposes Sysdig Monitor capabilities to LLMs, plus the cross-cutting SysQL tooling. New tools and functionalities will be added over time following semantic versioning. The goal is to provide a simple and easy-to-use interface for querying information from the Sysdig platform using LLMs.
+
+For Sysdig Secure-specific workflows, use the dedicated [@sysdig/secure-mcp-server](https://www.npmjs.com/package/@sysdig/secure-mcp-server).
 
 ## Quickstart Guide
 
 Get up and running with the Sysdig MCP Server quickly using our pre-built Docker image.
 
 1. **Get your API Token**:
-    Go to your Sysdig instance and navigate to **Settings > Sysdig Secure API** (or **Sysdig Monitor API**). Either a Sysdig Secure or Sysdig Monitor API token works. This token is required to authenticate requests to the Sysdig Platform (See the [Configuration](#configuration) section for more details).
+    Go to your Sysdig instance and navigate to **Settings > Sysdig Monitor API**. This token is required to authenticate requests to the Sysdig Platform (See the [Configuration](#configuration) section for more details).
 
 2. **Configure your MCP client**:
 
@@ -133,28 +142,6 @@ The server dynamically filters the available tools based on the permissions asso
 
 > **Note:** When a time window is provided, the underlying PromQL is wrapped in the aggregation appropriate for each tool (`avg_over_time`, `max_over_time`, `min_over_time`, `increase`, etc.) and evaluated at `end`. See [`internal/infra/mcp/tools/README.md`](./internal/infra/mcp/tools/README.md) for the per-tool aggregation table.
 
-### Sysdig Secure
-
-- **`list_runtime_events`**
-  - **Description**: List runtime security events from the last given hours, optionally filtered by severity level.
-  - **Required Permission**: `policy-events.read`
-  - **Sample Prompt**: "Show me high severity events from the last 2 hours in cluster1"
-
-- **`get_event_info`**
-  - **Description**: Retrieve detailed information for a specific security event by its ID.
-  - **Required Permission**: `policy-events.read`
-  - **Sample Prompt**: "Get full details for event ID 123abc"
-
-- **`get_event_process_tree`**
-  - **Description**: Retrieve the process tree for a specific event (if available).
-  - **Required Permission**: `policy-events.read`
-  - **Sample Prompt**: "Get the process tree for event ID abc123"
-
-- **`run_sysql`**
-  - **Description**: Execute a pre-written SysQL query directly (use only when user provides explicit query).
-  - **Required Permission**: `sage.exec`, `risks.read`
-  - **Sample Prompt**: "Run this query: MATCH CloudResource WHERE type = 'aws_s3_bucket' LIMIT 10"
-
 ### Sysdig Monitor & Sysdig Secure
 
 - **`generate_sysql`**
@@ -162,6 +149,11 @@ The server dynamically filters the available tools based on the permissions asso
   - **Required Permission**: `sage.exec`
   - **Sample Prompt**: "List top 10 pods by memory usage in the last hour"
   - **Note**: The `generate_sysql` tool currently does not work with Service Account tokens and will return a 500 error. For this tool, use an API token assigned to a regular user account.
+
+- **`run_sysql`**
+  - **Description**: Execute a pre-written SysQL query directly (use only when user provides explicit query).
+  - **Required Permission**: `sage.exec`, `risks.read`
+  - **Sample Prompt**: "Run this query: MATCH CloudResource WHERE type = 'aws_s3_bucket' LIMIT 10"
 
 ## Requirements
 - [Go](https://go.dev/doc/install) 1.26 or higher (if running without Docker).
@@ -224,7 +216,6 @@ To use the MCP server tools, your API token needs specific permissions in Sysdig
 | Permission           | Sysdig UI Permission Name                   |
 |----------------------|---------------------------------------------|
 | `metrics-data.read`  | Data Access Settings: "Metrics Data" (Read) |
-| `policy-events.read` | Threats: "Policy Events" (Read)             |
 | `risks.read`         | Risks: "Access to risk feature" (Read)      |
 | `sage.exec`          | SysQL: "AI Query Generation" (Exec)         |
